New profile 'remote' and filtering of resources in pom.xml
The new profile 'remote' renames the config_template.properties file into
config.properties before packaging. This is meant for remote deployments via
Jenkins. The variables to replace the placeholders in the template file have
to be given in Jenkins as command-line properties or via an additional profile.
In local deployments this has to be done by the user.

Formatted pom.xml.