Implements database updates when schema changed.
This includes a new property schema.version in the pom file, that is used to
write the current schema version in a file called 'schemaversion' (using maven
filtering). The Connection class compares that version to the one found in the
database (new table 'version') and updates the database if necessary. Updated
the SQL script to create the table 'version'.

The SQL script now uses a placeholder for the schema name that is set by
maven. The value should be set in a profile when building with Jenkins. The
default value 'JEngineV2' is defined as property in the pom file.