Refactor AbstractDatabaseDependentTest and AbstractTest
AbstractDatabaseDependentTest should be extended by all tests that
require the database. It clears the database before a test is run
and turns on the test mode of the ConnectionWrapper, so that it
connects to the test database.
AbstractTest is a duplicate of AbstractDatabaseDependentTest, but
extends JerseyTest. It is extended itself by all Tests that require
REST capabilites. FIXME: Find a way to merge those two.