Allows automatic execution of activities with *empty precondition* (see #21).
DOs are not considered yet!

When the control flow of an activity instance is enabled in its
TaskIncomingBehavior and the data flow is already enabled (i.e. it is in
state READY) and the activity instance is automatic, it begins execution.
So far, no data objects are selected to be worked on.

Only activities with at most one input and output set can be automatized.
SO FAR ONLY ACTIVITIES WITHOUT PRECONDITION START AUTOMATICALLY (see #50)
This is now checked. The method to set automatic execution flag was split
in two methods, 'allowAutomaticExecution()' and 'forbidAutomaticStart()'.