Change how WebServiceTaskInstance replaces variable expressions (#D.attr)
Now it does not try to replace with every attribute of selected data objects, but rather checks
first, whether a variable expression is present. It then looks up the referenced DO and attribute
and replaces the variable expression with the attribute value.
It works for states as well, #D is replaced with D's current state.
- Tests added as well
- extensive error logging