Fix boundary event locking bug (#159)
Boundary Events could not be executed if they want to lock the same data object as the attached activity. The enabling of the controlflow of the boundary event is handled differently than other Data Control Node Instances. If the boundary event and the attached activity have the same Condition in the Input Set although it might not be the same referenced Data Node, the Boundary Event prioritizes the selected Data Object of the attached Activity over an unlocked Data Object in the Case.