Automatic basic auth popup in browser.
When a frontend sends a request to Chimera and receives a 401
(unauthorized) response a popup should appear so the user can log in.
Because we use basic auth for authorization this commit adds a specific
header to the 401 response which tells the browser to automatically show
the standard login popup. To do so this commits modifies the
AuthorizationRequestFilter and adds the required header everytime a 401
error occurs.