Fix problems with database schema version and git
Because the mysql_schema_version.txt file containing your local running
database schema version was part of commits and could be pushed and
pulled maven detected no schema changes.
This commit adds mysql_schema_version.txt in the .gitignore. Also maven
now first checks for the existence of the file and creates it when
needed.